+++ CHANGELOG.md
@@ -9,4 +9,5 @@
 - New project structure
 - Added S3 based optimization cache support
 - Add support for TED playlists
-- Add support to filter videos available in a specific language
+- Add support to filter videos available in a specific language based on audio and subtitles
+- Add option to choose subtitle languages